DEV: Rename chat plugin service worker file for clarity
Rename `plugins/chat/assets/javascripts/service-worker.js` to
`chat-service-worker-extensions.js` so it doesn't share a filename
with the core service worker (`app/views/static/service-worker.js.erb`).
The new name better reflects its role: source code that's concatenated
into the core service worker via `register_service_worker`, not a
standalone worker file.
