+++ plugins/chat/assets/javascripts/chat-service-worker-extensions.js
